+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## [3.2.1] - 2021-03-10
+
+### Changed
+
+- The "Archives with dropdown" widget wraps its list of archive links in a `<ul>`
+
 
 ## [3.2.0] - 2021-02-16